+++ crates/python/src/qpu/isa.rs
@@ -56,13 +56,13 @@ py_wrap_simple_enum! {
 
 py_wrap_data_struct! {
     PyEdge(Edge) as "Edge" {
-        node_ids: Vec<i32> => Py<PyList>
+        node_ids: Vec<i64> => Py<PyList>
     }
 }
 
 py_wrap_data_struct! {
     PyNode(Node) as "Node" {
-        node_id: i32 => Py<PyInt>
+        node_id: i64 => Py<PyInt>
     }
 }
 
@@ -78,7 +78,7 @@ py_wrap_data_struct! {
     PyCharacteristic(Characteristic) as "Characteristic" {
         error: Option<f64> => Option<Py<PyFloat>>,
         name: String => Py<PyString>,
-        node_ids: Option<Vec<i32>> => Option<Py<PyList>>,
+        node_ids: Option<Vec<i64>> => Option<Py<PyList>>,
         parameter_values: Option<Vec<f64>> => Option<Py<PyList>>,
         timestamp: String => Py<PyString>,
         value: f64 => Py<PyFloat>
@@ -94,7 +94,7 @@ py_wrap_data_struct! {
 py_wrap_data_struct! {
     PyOperationSite(OperationSite) as "OperationSite" {
         characteristics: Vec<Characteristic> => Vec<PyCharacteristic>,
-        node_ids: Vec<i32> => Py<PyList>
+        node_ids: Vec<i64> => Py<PyList>
     }
 }
 
@@ -102,7 +102,7 @@ py_wrap_data_struct! {
     PyOperation(Operation) as "Operation" {
         characteristics: Vec<Characteristic> => Vec<PyCharacteristic>,
         name: String => Py<PyString>,
-        node_count: Option<i32> => Option<Py<PyInt>>,
+        node_count: Option<i64> => Option<Py<PyInt>>,
         parameters: Vec<Parameter> => Vec<PyParameter>,
         sites: Vec<OperationSite> => Vec<PyOperationSite>
     }
